continue refactory: insertItemForm
