+++ Views/src/themes/Vibrant/Components/SingleItem.jsx
@@ -4,39 +4,34 @@ import './singleItem.css';
 
 export default function SingleItem({
   mode = 'display',
-  Record = null, 
+  Record = {}, 
   ActionButtonsComponent = null,
   InfoComponent = null, 
   PlaceHolders = null 
 }) {
 
-  const {allergens,items,inHowManyReceipts, name,price,id,total,date,quantity} = Record || {};
+  const {allergens,items,inHowManyReceipts, name,price,id,total,date,quantity} = Record;
   const infoData = {allergens, items, inHowManyReceipts};
   const {actionProps, infoProps} = getComponentProps(mode, Record, infoData)
 
     
   return(
       <li className="sngl-item">
-
-        {InfoComponent &&
-          <div className="extra" >
-              <InfoComponent {...infoProps} /> 
-          </div> }
+        {InfoComponent && <InfoComponent {...infoProps} />}
 
         {Record &&
         <>
           <span className="first-record">{name ? name : date ? id +  " ~ " + date : ""}</span> 
           <span className="second-record">{price ? price + " €" : total ? total + " €" : quantity } </span>
-        </>
-      }
+        </>}
         
       {ActionButtonsComponent && <ActionButtonsComponent {...actionProps}/> }
+
       {PlaceHolders && PlaceHolders.map((label,i) =>{
         return(
           <span key={i} className={`label-${i+1}`}>{label}</span>
-          )
-        }
-      )}
+        )})}
       </li>
   )
 }
+
