--- Views/src/themes/Vibrant/tableComponents/Table/index.jsx
@@ -1,1 +1,0 @@
-export { default } from './Table';
