--- Views/src/themes/Vibrant/tableComponents/TableControls/TableControls.css
@@ -1,28 +1,0 @@
-.table__controls {
-  display: flex;
-  align-items: center;
-  justify-content: end;
-  height: 4rem;
-  position: relative;
-  width: 100%;
-  z-index: 1;
-}
-
-.table__controls-inner {
-  display: flex;
-  width: 100%;
-  justify-content: space-between;
-}
-
-.table__controls-inner-wrapper {
-  display: flex;
-}
-
-.check-icon-wrapper {
-  width: 30px;
-  height: 20px;
-  display: inline-flex;
-  align-items: center;
-  justify-content: center;
-  flex-shrink: 0;
-}
