--- Views/src/themes/Vibrant/tableComponents/TableControls/index.jsx
@@ -1,1 +1,0 @@
-export { default } from './TableControls';
