--- Views/src/themes/Vibrant/tableComponents/TableGroup/TableGroup.css
@@ -1,5 +1,0 @@
-.tablegroup {
-  display: flex;
-  width: 100%;
-  height: 75vh;
-}
