--- Views/src/themes/Vibrant/tableComponents/TableGroup/index.jsx
@@ -1,1 +1,0 @@
-export { default, useTableGroupContext } from './TableGroup';
