--- Views/src/themes/Vibrant/tableComponents/TableSection/TableSection.css
@@ -1,6 +1,0 @@
-.table__section {
-  display: flex;
-  flex-direction: column;
-  height: 100%;
-  padding-inline: 2rem;
-}
