--- Views/src/themes/Vibrant/tableComponents/TableSection/TableSection.jsx
@@ -1,11 +1,0 @@
-import './TableSection.css';
-
-function TableSection({ children, ...props }) {
-  return (
-    <div className="table__section" {...props}>
-      {children}
-    </div>
-  );
-}
-
-export default TableSection;
