--- Views/src/themes/Vibrant/tableComponents/TableSection/index.jsx
@@ -1,1 +1,0 @@
-export { default } from './TableSection';
